+++ CHANGELOG.rst
@@ -7,7 +7,8 @@ Breaking changes
 ~~~~~~~~~~~~~~~~
 - The radial coordinate input to `psipy.tracing.FortranTracer.trace` is now required to be
   a `~astropy.units.Quantity` (with length units) instead of a plain numpy array.
-
+- The radial coordinate input to the sphere methods on `psipy.visualization.pyvista.MASPlotter` is now required to be
+  a `~astropy.units.Quantity` (with length units) instead of a plain numpy array.
 Bug fixes
 ~~~~~~~~~
 - Fixed the in-situ sampling example.